# src/openapi.yaml
openapi: "3.0.3"
info:
  title: ScoutOff API
  description: >
    Backend API for ScoutOff — a decentralised football scouting platform
    built on the Stellar blockchain (Soroban smart contracts).
  version: "1.0.0"
  contact:
    name: ScoutOff Engineering
servers:
  - url: /api
    description: Stable API (version-less alias)
  - url: /api/v1
    description: Versioned API
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: >-
        Long-lived scout API key (alternative to a Bearer JWT). Restricted keys
        may only perform operations covered by their granted scopes (#1019).
        Legacy keys without a scope list are unrestricted.
  parameters:
...
    walletParam:
      name: wallet
      in: path
      required: true
      schema:
        type: string
        minLength: 56
        maxLength: 56
      description: Stellar public key (G…)
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [success, error]
      properties:
        success:
          type: boolean
          enum: [false]
        error:
          type: string
        code:
          type: string

    # ── Player ────────────────────────────────────────────────────────────────
...
  responses:
    Unauthorized:
      description: Missing or invalid JWT
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    Forbidden:
      description: Insufficient role or wallet mismatch
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
paths:
...
  /scouts/{wallet}/recommendations:
    get:
      tags: [scouts]
      summary: Get personalised player recommendations for a scout
      operationId: getScoutRecommendations
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/walletParam"
      responses:
        "200":
          description: Recommendation list
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /admin/events/export:
    get:
      tags: [admin]
      summary: Export events as CSV
      operationId: exportEvents
      security:
        - bearerAuth: []
      parameters:
        - name: startDate
          in: query
          schema:
            type: string
            format: date-time
        - name: endDate
          in: query
          schema:
            type: string
            format: date-time
        - name: eventType
          in: query
          schema:
            type: string
      responses:
        "200":
          description: CSV download
          content:
            text/csv:
              schema:
                type: string
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...